Move all the files to live under paths
