+++ src/Table/Table.story.tsx
@@ -2,8 +2,8 @@
 import { jsx } from "@emotion/core";
 import React from "react";
 import { storiesOf } from "@storybook/react";
-import { colors } from ".";
-import * as typography from "./typography";
+import * as colors from "../colors";
+import * as typography from "../typography";
 import { Table } from "./Table";
 import { withKnobs, select } from "@storybook/addon-knobs/react";
