+++ src/Table/Table.tsx
@@ -1,8 +1,8 @@
 /** @jsx jsx */
 import { jsx } from "@emotion/core";
 import React from "react";
-import * as typography from "./typography";
-import * as colors from "./colors";
+import * as typography from "../typography";
+import * as colors from "../colors";
 
 interface Props<RowShape> {
   /**
