--- src/index.ts
@@ -1,9 +1,0 @@
-import * as colors from "./colors";
-import * as fonts from "./fonts";
-
-export { colors };
-export { fonts };
-
-export { Table } from "./Table";
-export { Button } from "./Button";
-export { Modal } from "./Modal";
