+++ src/typography/index.ts
@@ -1,5 +1,1 @@
-import * as base from "./base";
-import * as mono from "./mono";
-
-export { base };
-export { mono };
+export * from "./typography";
