+++ src/typography/typography.ts
@@ -1,0 +1,5 @@
+import * as base from "./base";
+import * as mono from "./mono";
+
+export { base };
+export { mono };
