+++ tsconfig.json
@@ -1,11 +1,16 @@
 {
   "extends": "./tsconfig.base.json",
-  "exclude": ["./node_modules"],
+  "exclude": [
+    "./node_modules",
+    "*.spec.*",
+    "*.story.*",
+  ],
   "references": [
     {
       "path": "./icons"
     }
   ],
-  "files": ["src/index.ts", "src/fonts.ts", "src/Button.tsx"],
-  "include": ["./src/typography/*.ts", "./src/colors/**/*.ts"]
+  "include": [
+    "./src/**/*"
+  ]
 }
